move DailyStatus
The `DailyStatus` class remained in a separate folder from when we used self-written data classes. It has now been moved to the basic `model` folder.
