+++ src/app/util.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { AppConfigService } from './app-config.service';
 import dayjs from 'dayjs';
-import { DailyStatus } from './model/frontend/daily-status';
+import { DailyStatus } from './model/daily-status';
 import { UserSettingsService } from './user-settings.service';
 import { Incident, IncidentUpdate } from '../external/lib/status-page-api/angular-client';
 import { DataService } from './data.service';
